+++ src/camera.cpp
@@ -1,7 +1,7 @@
 #include "camera.h"
 
 
-Camera::Camera(int id, double fx, double fy, double cx, double cy, double fb)
+Camera::Camera(int id, float fx, float fy, float cx, float cy, float fb)
 :_id(id),_fx(fx),_fy(fy),_cx(cx),_cy(cy),_fb(fb)
 {
